moving configs into configs/
