+++ src/covid19sim/models/run.py
@@ -4,7 +4,7 @@ import functools
 from joblib import Parallel, delayed
 
 from covid19sim.server_utils import InferenceClient, InferenceWorker
-from covid19sim import config
+from covid19sim.configs import config
 from ctt.inference.infer import InferenceEngine
