+++ src/covid19sim/simulator.py
@@ -4,7 +4,7 @@ from covid19sim.utils import _normalize_scores, _get_random_sex, _get_covid_prog
      _get_preexisting_conditions, _draw_random_discreet_gaussian, _sample_viral_load_piecewise, \
      _get_cold_progression, _get_flu_progression, _get_allergy_progression, _get_get_really_sick
 from covid19sim.base import *
-from covid19sim.constants import BIG_NUMBER, TICK_MINUTE
+from covid19sim.configs.constants import BIG_NUMBER, TICK_MINUTE
 
 # if env.exp_config.COLLECT_LOGS is False:
 #     Event = DummyEvent
